separate node and browser code into separate exports
