+++ packages/app-config/src/node/index.ts
@@ -1,0 +1,2 @@
+export * from '../index';
+export * from './module-federation';
